skip valid authorizations
Authorizations that already have a status of "valid" do not need
to be validated again (until the authorization expires).
Skipping these authorizations makes it possible to obtain certificates
for domains for which a dns-01 validation is still valid, as
acme_tiny is unable to re-validate them. (The alternative would be
to wait up to a month until the authorization expires, or use a
different acme client.)

(cherry picked from commit c1a57b1ab7439d02b67c73c3d14e27ce8c9ab515)